Moved extra code from SceneBoundariesEntityHandler into SBCDebugMode; removed SBEH; Added comments with questions
