--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/SceneBoundariesController/SceneBoundariesEntityHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23627fae21c6de243b3b4b4d82ba6926
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
