+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/SceneController.cs
@@ -47,7 +47,7 @@ namespace DCL
             Debug.unityLogger.logEnabled = false;
 #endif
 
-            InitializeSceneBoundariesChecker();
+            InitializeSceneBoundariesChecker(isDebugMode);
 
             RenderProfileManifest.i.Initialize();
             Environment.i.Initialize(this, this);
@@ -220,99 +220,99 @@ namespace DCL
                 switch (method)
                 {
                     case MessagingTypes.ENTITY_CREATE:
-                    {
-                        if (msgObject.payload is Protocol.CreateEntity payload)
-                            scene.CreateEntity(payload.entityId);
+                        {
+                            if (msgObject.payload is Protocol.CreateEntity payload)
+                                scene.CreateEntity(payload.entityId);
 
-                        break;
-                    }
+                            break;
+                        }
                     case MessagingTypes.ENTITY_REPARENT:
-                    {
-                        if (msgObject.payload is Protocol.SetEntityParent payload)
-                            scene.SetEntityParent(payload.entityId, payload.parentId);
+                        {
+                            if (msgObject.payload is Protocol.SetEntityParent payload)
+                                scene.SetEntityParent(payload.entityId, payload.parentId);
 
-                        break;
-                    }
+                            break;
+                        }
 
                     case MessagingTypes.ENTITY_COMPONENT_CREATE_OR_UPDATE:
-                    {
-                        if (msgObject.payload is Protocol.EntityComponentCreateOrUpdate payload)
-                            scene.EntityComponentCreateOrUpdate(payload.entityId, (CLASS_ID_COMPONENT) payload.classId, payload.json, out yieldInstruction);
+                        {
+                            if (msgObject.payload is Protocol.EntityComponentCreateOrUpdate payload)
+                                scene.EntityComponentCreateOrUpdate(payload.entityId, (CLASS_ID_COMPONENT)payload.classId, payload.json, out yieldInstruction);
 
-                        break;
-                    }
+                            break;
+                        }
 
                     case MessagingTypes.ENTITY_COMPONENT_DESTROY:
-                    {
-                        if (msgObject.payload is Protocol.EntityComponentDestroy payload)
-                            scene.EntityComponentRemove(payload.entityId, payload.name);
+                        {
+                            if (msgObject.payload is Protocol.EntityComponentDestroy payload)
+                                scene.EntityComponentRemove(payload.entityId, payload.name);
 
-                        break;
-                    }
+                            break;
+                        }
 
                     case MessagingTypes.SHARED_COMPONENT_ATTACH:
-                    {
-                        if (msgObject.payload is Protocol.SharedComponentAttach payload)
-                            scene.SharedComponentAttach(payload.entityId, payload.id);
+                        {
+                            if (msgObject.payload is Protocol.SharedComponentAttach payload)
+                                scene.SharedComponentAttach(payload.entityId, payload.id);
 
-                        break;
-                    }
+                            break;
+                        }
 
                     case MessagingTypes.SHARED_COMPONENT_CREATE:
-                    {
-                        if (msgObject.payload is Protocol.SharedComponentCreate payload)
-                            scene.SharedComponentCreate(payload.id, payload.classId);
+                        {
+                            if (msgObject.payload is Protocol.SharedComponentCreate payload)
+                                scene.SharedComponentCreate(payload.id, payload.classId);
 
-                        break;
-                    }
+                            break;
+                        }
 
                     case MessagingTypes.SHARED_COMPONENT_DISPOSE:
-                    {
-                        if (msgObject.payload is Protocol.SharedComponentDispose payload)
-                            scene.SharedComponentDispose(payload.id);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is Protocol.SharedComponentDispose payload)
+                                scene.SharedComponentDispose(payload.id);
+                            break;
+                        }
 
                     case MessagingTypes.SHARED_COMPONENT_UPDATE:
-                    {
-                        if (msgObject.payload is Protocol.SharedComponentUpdate payload)
-                            scene.SharedComponentUpdate(payload.componentId, payload.json, out yieldInstruction);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is Protocol.SharedComponentUpdate payload)
+                                scene.SharedComponentUpdate(payload.componentId, payload.json, out yieldInstruction);
+                            break;
+                        }
 
                     case MessagingTypes.ENTITY_DESTROY:
-                    {
-                        if (msgObject.payload is Protocol.RemoveEntity payload)
-                            scene.RemoveEntity(payload.entityId);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is Protocol.RemoveEntity payload)
+                                scene.RemoveEntity(payload.entityId);
+                            break;
+                        }
 
                     case MessagingTypes.INIT_DONE:
-                    {
-                        scene.SetInitMessagesDone();
-                        break;
-                    }
+                        {
+                            scene.SetInitMessagesDone();
+                            break;
+                        }
 
                     case MessagingTypes.QUERY:
-                    {
-                        if (msgObject.payload is QueryMessage queryMessage)
-                            ParseQuery(queryMessage.payload, scene.sceneData.id);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is QueryMessage queryMessage)
+                                ParseQuery(queryMessage.payload, scene.sceneData.id);
+                            break;
+                        }
 
                     case MessagingTypes.OPEN_EXTERNAL_URL:
-                    {
-                        if (msgObject.payload is Protocol.OpenExternalUrl payload)
-                            OnOpenExternalUrlRequest?.Invoke(scene, payload.url);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is Protocol.OpenExternalUrl payload)
+                                OnOpenExternalUrlRequest?.Invoke(scene, payload.url);
+                            break;
+                        }
 
                     case MessagingTypes.OPEN_NFT_DIALOG:
-                    {
-                        if (msgObject.payload is Protocol.OpenNftDialog payload)
-                            OnOpenNFTDialogRequest?.Invoke(payload.contactAddress, payload.tokenId, payload.comment);
-                        break;
-                    }
+                        {
+                            if (msgObject.payload is Protocol.OpenNftDialog payload)
+                                OnOpenNFTDialogRequest?.Invoke(payload.contactAddress, payload.tokenId, payload.comment);
+                            break;
+                        }
 
                     default:
                         Debug.LogError($"Unknown method {method}");
@@ -359,7 +359,7 @@ namespace DCL
 
         private string SendSceneMessage(string payload, bool enqueue)
         {
-            string[] chunks = payload.Split(new char[] {'\n'}, StringSplitOptions.RemoveEmptyEntries);
+            string[] chunks = payload.Split(new char[] { '\n' }, StringSplitOptions.RemoveEmptyEntries);
             int count = chunks.Length;
             string lastBusId = null;
 
@@ -467,7 +467,7 @@ namespace DCL
 
             if (data.parcels == null)
             {
-                data.parcels = new Vector2Int[] {data.basePosition};
+                data.parcels = new Vector2Int[] { data.basePosition };
             }
 
             if (string.IsNullOrEmpty(data.id))
@@ -566,25 +566,39 @@ namespace DCL
             return solvedPosition;
         }
 
-        void InitializeSceneBoundariesChecker()
+        // TODO: We should call this with the pertinent param when we toggle the builder in-world
+        void InitializeSceneBoundariesChecker(bool debugMode)
         {
             if (!useBoundariesChecker) return;
 
             if (boundariesChecker != null)
                 boundariesChecker.Stop();
 
+            if (debugMode)
+            {
+                boundariesChecker = new SceneBoundariesDebugModeChecker();
+                boundariesChecker.timeBetweenChecks = 0f;
+            }
+            else
+            {
+                boundariesChecker = new SceneBoundariesChecker();
+            }
+
+            // BUILDER in-world
+            /*
             boundariesChecker = new SceneBoundariesDebugModeChecker();
             if (isDebugMode)
             {
                 boundariesChecker.SetDebugMode();
                 boundariesChecker.timeBetweenChecks = 0f;
             }
+            */
         }
 
         private void SetPositionDirty(DCLCharacterPosition character)
         {
-            var currentX = (int) Math.Floor(character.worldPosition.x / ParcelSettings.PARCEL_SIZE);
-            var currentY = (int) Math.Floor(character.worldPosition.z / ParcelSettings.PARCEL_SIZE);
+            var currentX = (int)Math.Floor(character.worldPosition.x / ParcelSettings.PARCEL_SIZE);
+            var currentY = (int)Math.Floor(character.worldPosition.z / ParcelSettings.PARCEL_SIZE);
 
             positionDirty = currentX != currentGridSceneCoordinate.x || currentY != currentGridSceneCoordinate.y;
 
@@ -740,7 +754,7 @@ namespace DCL
         public void UnloadScene(string sceneKey)
         {
             var queuedMessage = new MessagingBus.QueuedSceneMessage()
-                {type = MessagingBus.QueuedSceneMessage.Type.UNLOAD_PARCEL, message = sceneKey};
+            { type = MessagingBus.QueuedSceneMessage.Type.UNLOAD_PARCEL, message = sceneKey };
 
             OnMessageWillQueue?.Invoke(MessagingTypes.SCENE_DESTROY);
 
@@ -809,7 +823,7 @@ namespace DCL
         public void UpdateParcelScenes(string decentralandSceneJSON)
         {
             var queuedMessage = new MessagingBus.QueuedSceneMessage()
-                {type = MessagingBus.QueuedSceneMessage.Type.UPDATE_PARCEL, message = decentralandSceneJSON};
+            { type = MessagingBus.QueuedSceneMessage.Type.UPDATE_PARCEL, message = decentralandSceneJSON };
 
             OnMessageWillQueue?.Invoke(MessagingTypes.SCENE_UPDATE);
 
@@ -818,7 +832,7 @@ namespace DCL
 
         public void UnloadAllScenesQueued()
         {
-            var queuedMessage = new MessagingBus.QueuedSceneMessage() {type = MessagingBus.QueuedSceneMessage.Type.UNLOAD_SCENES};
+            var queuedMessage = new MessagingBus.QueuedSceneMessage() { type = MessagingBus.QueuedSceneMessage.Type.UNLOAD_SCENES };
 
             OnMessageWillQueue?.Invoke(MessagingTypes.SCENE_DESTROY);
 
@@ -935,7 +949,7 @@ namespace DCL
             isDebugMode = true;
             fpsPanel.SetActive(true);
 
-            InitializeSceneBoundariesChecker();
+            InitializeSceneBoundariesChecker(true);
 
             OnDebugModeSet?.Invoke();
